move function of summing score to UnitScoreBoard class
