+++ src/main/java/chess/domain/chess/unit/Bishop.java
@@ -8,7 +8,6 @@ import chess.domain.geometric.Vector;
 import java.util.List;
 
 public class Bishop extends Unit {
-    private static final double SCORE = UnitInformation.BISHOP.score();
     private static List<Direction> directions;
 
     static {
@@ -23,11 +22,6 @@ public class Bishop extends Unit {
     public boolean validateDirection(Vector another) {
         return directions.stream()
                 .anyMatch(direction -> direction.isParallelTo(another));
-    }
-
-    @Override
-    public double score() {
-        return SCORE;
     }
 
     @Override
