+++ src/main/java/chess/domain/chess/unit/Queen.java
@@ -11,8 +11,6 @@ import java.util.List;
 import static java.util.stream.Collectors.toList;
 
 public class Queen extends Unit {
-    private static final double SCORE = UnitInformation.QUEEN.score();
-
     private static List<Direction> directions;
 
     static {
@@ -29,12 +27,6 @@ public class Queen extends Unit {
         return directions.stream()
                 .anyMatch(direction -> direction.isParallelTo(another));
     }
-
-    @Override
-    public double score() {
-        return SCORE;
-    }
-
     @Override
     public String toString() {
         return getTeam().applyUpperCaseOrLowerCaseByTeam(UnitInformation.QUEEN.symbol());
