+++ src/main/java/chess/domain/chess/unit/Rook.java
@@ -8,8 +8,6 @@ import chess.domain.geometric.Vector;
 import java.util.List;
 
 public class Rook extends Unit {
-    private static final double SCORE = UnitInformation.ROOK.score();
-
     private static List<Direction> directions;
 
     static {
@@ -26,15 +24,6 @@ public class Rook extends Unit {
                 .anyMatch(vector -> vector.isParallelTo(another));
     }
 
-    @Override
-    public double score() {
-        return SCORE;
-    }
-
-    @Override
-    public String getName() {
-        return "룩";
-    }
 
     @Override
     public String toString() {
